Refactor/replace long with double in check remaining days (#32)
* fix: change remainingDays type from long to double

* refactor: use double for remainingDays in LowRemainingDaysAlertService

* feat: add DaysFormatter utility for remaining days display

* refactor: use DaysFormatter in LowRemainingDaysAlertService

* refactor: format remaining days in LowRemainingDaysAlertRequestedService

* refactor: use DaysFormatter in calendar template

+++ src/main/java/school/hei/asa/endpoint/event/model/LowRemainingDaysAlertRequested.java
@@ -18,7 +18,7 @@ import lombok.ToString;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class LowRemainingDaysAlertRequested extends PojaEvent {
   private String workerCode;
-  private long remainingDays;
+  private double remainingDays;
 
   @Override
   public Duration maxConsumerDuration() {
